Move SourceLocation and standardize use of debug::ColumnType
